@@ -7,13 +7,8 @@ from typing import cast
 from typing import Optional
 from typing import Union
 
-from lexicon.providers import linode
-from lexicon.providers import linode4
-
 from certbot import errors
-from certbot.plugins import dns_common
 from certbot.plugins import dns_common_lexicon
-from certbot.plugins.dns_common import CredentialsConfiguration
 
 logger = logging.getLogger(__name__)
 
@@ -21,7 +16,7 @@ API_KEY_URL = 'https://manager.linode.com/profile/api'
 API_KEY_URL_V4 = 'https://cloud.linode.com/profile/tokens'
 
 
-class Authenticator(dns_common.DNSAuthenticator):
+class Authenticator(dns_common_lexicon.LexiconDNSAuthenticator):
     """DNS Authenticator for Linode
 
     This Authenticator uses the Linode API to fulfill a dns-01 challenge.
@@ -31,7 +26,10 @@ class Authenticator(dns_common.DNSAuthenticator):
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
-        self.credentials: Optional[CredentialsConfiguration] = None
+        self._add_provider_option('key',
+                                  'API key for Linode account, '
+                                  f'obtained from {API_KEY_URL} or {API_KEY_URL_V4}',
+                                  'auth_token')
 
     @classmethod
     def add_parser_arguments(cls, add: Callable[..., None],
@@ -43,29 +41,13 @@ class Authenticator(dns_common.DNSAuthenticator):
         return 'This plugin configures a DNS TXT record to respond to a dns-01 challenge using ' + \
                'the Linode API.'
 
-    def _setup_credentials(self) -> None:
-        self.credentials = self._configure_credentials(
-            'credentials',
-            'Linode credentials INI file',
-            {
-                'key': 'API key for Linode account, obtained from {0} or {1}'
-                        .format(API_KEY_URL, API_KEY_URL_V4)
-            }
-        )
+    @property
+    def _provider_name(self) -> str:
+        if not hasattr(self, '_credentials'):  # pragma: no cover
+            self._setup_credentials()
 
-    def _perform(self, domain: str, validation_name: str, validation: str) -> None:
-        self._get_linode_client().add_txt_record(domain, validation_name, validation)
-
-    def _cleanup(self, domain: str, validation_name: str, validation: str) -> None:
-        self._get_linode_client().del_txt_record(domain, validation_name, validation)
-
-    def _get_linode_client(self) -> '_LinodeLexiconClient':
-        if not self.credentials:  # pragma: no cover
-            raise errors.Error("Plugin has not been prepared.")
-        api_key = cast(str, self.credentials.conf('key'))
-        api_version: Optional[Union[str, int]] = self.credentials.conf('version')
-        if api_version == '':
-            api_version = None
+        api_key = cast(str, self._credentials.conf('key'))
+        api_version: Optional[Union[str, int]] = self._credentials.conf('version')
 
         if not api_version:
             api_version = 3
@@ -78,34 +60,19 @@ class Authenticator(dns_common.DNSAuthenticator):
         else:
             api_version = int(api_version)
 
-        return _LinodeLexiconClient(api_key, api_version)
-
-
-class _LinodeLexiconClient(dns_common_lexicon.LexiconClient):
-    """
-    Encapsulates all communication with the Linode API.
-    """
-
-    def __init__(self, api_key: str, api_version: int) -> None:
-        super().__init__()
-
-        self.api_version = api_version
-
         if api_version == 3:
-            config = dns_common_lexicon.build_lexicon_config('linode', {}, {
-                'auth_token': api_key,
-            })
-
-            self.provider = linode.Provider(config)
+            return 'linode'
         elif api_version == 4:
-            config = dns_common_lexicon.build_lexicon_config('linode4', {}, {
-                'auth_token': api_key,
-            })
+            return 'linode4'
 
-            self.provider = linode4.Provider(config)
-        else:
-            raise errors.PluginError('Invalid api version specified: {0}. (Supported: 3, 4)'
-                                     .format(api_version))
+        raise errors.PluginError(f'Invalid api version specified: {api_version}. (Supported: 3, 4)')
+
+    def _setup_credentials(self) -> None:
+        self._credentials = self._configure_credentials(
+            key='credentials',
+            label='Credentials INI file for linode DNS authenticator',
+            required_variables={item[0]: item[1] for item in self._provider_options},
+        )
 
     def _handle_general_error(self, e: Exception, domain_name: str) -> Optional[errors.PluginError]:
         if not str(e).startswith('Domain not found'):
