@@ -2,23 +2,19 @@
 import logging
 from typing import Any
 from typing import Callable
-from typing import cast
 from typing import Optional
 
-from lexicon.providers import dnsmadeeasy
 from requests import HTTPError
 
 from certbot import errors
-from certbot.plugins import dns_common
 from certbot.plugins import dns_common_lexicon
-from certbot.plugins.dns_common import CredentialsConfiguration
 
 logger = logging.getLogger(__name__)
 
 ACCOUNT_URL = 'https://cp.dnsmadeeasy.com/account/info'
 
 
-class Authenticator(dns_common.DNSAuthenticator):
+class Authenticator(dns_common_lexicon.LexiconDNSAuthenticator):
     """DNS Authenticator for DNS Made Easy
 
     This Authenticator uses the DNS Made Easy API to fulfill a dns-01 challenge.
@@ -26,11 +22,17 @@ class Authenticator(dns_common.DNSAuthenticator):
 
     description = ('Obtain certificates using a DNS TXT record (if you are using DNS Made Easy for '
                    'DNS).')
-    ttl = 60
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
-        self.credentials: Optional[CredentialsConfiguration] = None
+        self._add_provider_option('api-key',
+                                  'API key for DNS Made Easy account, '
+                                  f'obtained from {ACCOUNT_URL}',
+                                  'auth_username')
+        self._add_provider_option('secret-key',
+                                  'Secret key for DNS Made Easy account, '
+                                  f'obtained from {ACCOUNT_URL}',
+                                  'auth_token')
 
     @classmethod
     def add_parser_arguments(cls, add: Callable[..., None],
@@ -42,48 +44,9 @@ class Authenticator(dns_common.DNSAuthenticator):
         return 'This plugin configures a DNS TXT record to respond to a dns-01 challenge using ' + \
                'the DNS Made Easy API.'
 
-    def _setup_credentials(self) -> None:
-        self.credentials = self._configure_credentials(
-            'credentials',
-            'DNS Made Easy credentials INI file',
-            {
-                'api-key': 'API key for DNS Made Easy account, obtained from {0}'
-                           .format(ACCOUNT_URL),
-                'secret-key': 'Secret key for DNS Made Easy account, obtained from {0}'
-                              .format(ACCOUNT_URL)
-            }
-        )
-
-    def _perform(self, domain: str, validation_name: str, validation: str) -> None:
-        self._get_dnsmadeeasy_client().add_txt_record(domain, validation_name, validation)
-
-    def _cleanup(self, domain: str, validation_name: str, validation: str) -> None:
-        self._get_dnsmadeeasy_client().del_txt_record(domain, validation_name, validation)
-
-    def _get_dnsmadeeasy_client(self) -> "_DNSMadeEasyLexiconClient":
-        if not self.credentials:  # pragma: no cover
-            raise errors.Error("Plugin has not been prepared.")
-        return _DNSMadeEasyLexiconClient(cast(str, self.credentials.conf('api-key')),
-                                         cast(str, self.credentials.conf('secret-key')),
-                                         self.ttl)
-
-
-class _DNSMadeEasyLexiconClient(dns_common_lexicon.LexiconClient):
-    """
-    Encapsulates all communication with the DNS Made Easy via Lexicon.
-    """
-
-    def __init__(self, api_key: str, secret_key: str, ttl: int) -> None:
-        super().__init__()
-
-        config = dns_common_lexicon.build_lexicon_config('dnsmadeeasy', {
-            'ttl': ttl,
-        }, {
-            'auth_username': api_key,
-            'auth_token': secret_key,
-        })
-
-        self.provider = dnsmadeeasy.Provider(config)
+    @property
+    def _provider_name(self) -> str:
+        return 'dnsmadeeasy'
 
     def _handle_http_error(self, e: HTTPError, domain_name: str) -> Optional[errors.PluginError]:
         if domain_name in str(e) and str(e).startswith('404 Client Error: Not Found for url:'):
