@@ -1,10 +1,9 @@
 """Tests for certbot_dns_luadns._internal.dns_luadns."""
-
 import sys
-import unittest
 from unittest import mock
 
 import pytest
+from requests import Response
 from requests.exceptions import HTTPError
 
 from certbot.compat import os
@@ -17,7 +16,9 @@ TOKEN = 'foo'
 
 
 class AuthenticatorTest(test_util.TempDirTestCase,
-                        dns_test_common_lexicon.BaseLexiconAuthenticatorTest):
+                        dns_test_common_lexicon.BaseLexiconDNSAuthenticatorTest):
+
+    LOGIN_ERROR = HTTPError("401 Client Error: Unauthorized for url: ...", response=Response())
 
     def setUp(self):
         super().setUp()
@@ -31,23 +32,6 @@ class AuthenticatorTest(test_util.TempDirTestCase,
                                      luadns_propagation_seconds=0)  # don't wait during tests
 
         self.auth = Authenticator(self.config, "luadns")
-
-        self.mock_client = mock.MagicMock()
-        # _get_luadns_client | pylint: disable=protected-access
-        self.auth._get_luadns_client = mock.MagicMock(return_value=self.mock_client)
-
-
-class LuaDNSLexiconClientTest(unittest.TestCase, dns_test_common_lexicon.BaseLexiconClientTest):
-
-    LOGIN_ERROR = HTTPError("401 Client Error: Unauthorized for url: ...")
-
-    def setUp(self):
-        from certbot_dns_luadns._internal.dns_luadns import _LuaDNSLexiconClient
-
-        self.client = _LuaDNSLexiconClient(EMAIL, TOKEN, 0)
-
-        self.provider_mock = mock.MagicMock()
-        self.client.provider = self.provider_mock
 
 
 if __name__ == "__main__":
