@@ -2,34 +2,31 @@
 import logging
 from typing import Any
 from typing import Callable
-from typing import cast
 from typing import Optional
 
-from lexicon.providers import nsone
 from requests import HTTPError
 
 from certbot import errors
-from certbot.plugins import dns_common
 from certbot.plugins import dns_common_lexicon
-from certbot.plugins.dns_common import CredentialsConfiguration
 
 logger = logging.getLogger(__name__)
 
 ACCOUNT_URL = 'https://my.nsone.net/#/account/settings'
 
 
-class Authenticator(dns_common.DNSAuthenticator):
-    """DNS Authenticator for NS1
-
+class Authenticator(dns_common_lexicon.LexiconDNSAuthenticator):
+    """
+    DNS Authenticator for NS1
     This Authenticator uses the NS1 API to fulfill a dns-01 challenge.
     """
 
     description = 'Obtain certificates using a DNS TXT record (if you are using NS1 for DNS).'
-    ttl = 60
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
-        self.credentials: Optional[CredentialsConfiguration] = None
+        self._add_provider_option('api-key',
+                                  f'API key for NS1 API, obtained from {ACCOUNT_URL}',
+                                  'auth_token')
 
     @classmethod
     def add_parser_arguments(cls, add: Callable[..., None],
@@ -41,42 +38,9 @@ class Authenticator(dns_common.DNSAuthenticator):
         return 'This plugin configures a DNS TXT record to respond to a dns-01 challenge using ' + \
                'the NS1 API.'
 
-    def _setup_credentials(self) -> None:
-        self.credentials = self._configure_credentials(
-            'credentials',
-            'NS1 credentials file',
-            {
-                'api-key': 'API key for NS1 API, obtained from {0}'.format(ACCOUNT_URL)
-            }
-        )
-
-    def _perform(self, domain: str, validation_name: str, validation: str) -> None:
-        self._get_nsone_client().add_txt_record(domain, validation_name, validation)
-
-    def _cleanup(self, domain: str, validation_name: str, validation: str) -> None:
-        self._get_nsone_client().del_txt_record(domain, validation_name, validation)
-
-    def _get_nsone_client(self) -> "_NS1LexiconClient":
-        if not self.credentials:  # pragma: no cover
-            raise errors.Error("Plugin has not been prepared.")
-        return _NS1LexiconClient(cast(str, self.credentials.conf('api-key')), self.ttl)
-
-
-class _NS1LexiconClient(dns_common_lexicon.LexiconClient):
-    """
-    Encapsulates all communication with the NS1 via Lexicon.
-    """
-
-    def __init__(self, api_key: str, ttl: int) -> None:
-        super().__init__()
-
-        config = dns_common_lexicon.build_lexicon_config('nsone', {
-            'ttl': ttl,
-        }, {
-            'auth_token': api_key,
-        })
-
-        self.provider = nsone.Provider(config)
+    @property
+    def _provider_name(self) -> str:
+        return 'nsone'
 
     def _handle_http_error(self, e: HTTPError, domain_name: str) -> Optional[errors.PluginError]:
         if domain_name in str(e) and (str(e).startswith('404 Client Error: Not Found for url:') or
