@@ -1,10 +1,9 @@
 """Tests for certbot_dns_ovh._internal.dns_ovh."""
-
-import sys
-import unittest
 from unittest import mock
+import sys
 
 import pytest
+from requests import Response
 from requests.exceptions import HTTPError
 
 from certbot.compat import os
@@ -19,8 +18,10 @@ CONSUMER_KEY = 'spam'
 
 
 class AuthenticatorTest(test_util.TempDirTestCase,
-                        dns_test_common_lexicon.BaseLexiconAuthenticatorTest):
+                        dns_test_common_lexicon.BaseLexiconDNSAuthenticatorTest):
 
+    DOMAIN_NOT_FOUND = Exception('Domain example.com not found')
+    LOGIN_ERROR = HTTPError('403 Client Error: Forbidden for url: https://eu.api.ovh.com/1.0/...', response=Response())
     def setUp(self):
         super().setUp()
 
@@ -38,26 +39,7 @@ class AuthenticatorTest(test_util.TempDirTestCase,
         self.config = mock.MagicMock(ovh_credentials=path,
                                      ovh_propagation_seconds=0)  # don't wait during tests
 
-        self.auth = Authenticator(self.config, "ovh")
-
-        self.mock_client = mock.MagicMock()
-        # _get_ovh_client | pylint: disable=protected-access
-        self.auth._get_ovh_client = mock.MagicMock(return_value=self.mock_client)
-
-
-class OVHLexiconClientTest(unittest.TestCase, dns_test_common_lexicon.BaseLexiconClientTest):
-    DOMAIN_NOT_FOUND = Exception('Domain example.com not found')
-    LOGIN_ERROR = HTTPError('403 Client Error: Forbidden for url: https://eu.api.ovh.com/1.0/...')
-
-    def setUp(self):
-        from certbot_dns_ovh._internal.dns_ovh import _OVHLexiconClient
-
-        self.client = _OVHLexiconClient(
-            ENDPOINT, APPLICATION_KEY, APPLICATION_SECRET, CONSUMER_KEY, 0
-        )
-
-        self.provider_mock = mock.MagicMock()
-        self.client.provider = self.provider_mock
+        self.auth = Authenticator(self.config, 'ovh')
 
 
 if __name__ == "__main__":
