@@ -117,7 +117,7 @@ def _check_version(version_str):
 
     version = (int(search.group(1)), int(search.group(2)))
 
-    if version >= (3, 7):
+    if version >= (3, 8):
         return True
 
     print('Incompatible python version for Certbot found: {0}'.format(version_str))
